@@ -2,7 +2,7 @@
  * @Author: Cao Shixin
  * @Date: 2020-12-31 04:27:56
  * @LastEditors: Cao Shixin
- * @LastEditTime: 2021-01-05 11:47:53
+ * @LastEditTime: 2021-03-25 17:06:16
  * @Description: 
  */
 
@@ -12,15 +12,15 @@
 class IosDeviceInfo {
   /// IOS device info class.
   IosDeviceInfo({
-    this.name,
-    this.systemName,
-    this.systemVersion,
-    this.model,
-    this.localizedModel,
-    this.identifierForVendor,
-    this.isPhysicalDevice,
-    this.utsname,
-    this.storage,
+    required this.name,
+    required this.systemName,
+    required this.systemVersion,
+    required this.model,
+    required this.localizedModel,
+    required this.identifierForVendor,
+    required this.isPhysicalDevice,
+    required this.utsname,
+    required this.storage,
   });
 
   /// Device name.
@@ -53,15 +53,21 @@ class IosDeviceInfo {
   /// Deserializes from the map message received from [_kChannel].
   static IosDeviceInfo fromMap(Map<String, dynamic> map) {
     return IosDeviceInfo(
-      name: map['name'],
-      systemName: map['systemName'],
-      systemVersion: map['systemVersion'],
-      model: map['model'],
-      localizedModel: map['localizedModel'],
-      identifierForVendor: map['identifierForVendor'],
-      isPhysicalDevice: map['isPhysicalDevice'] == 'true',
-      utsname: IosUtsname._fromMap(map['utsname'].cast<String, dynamic>()),
-      storage: Storage._fromMap(map['storage'].cast<String, dynamic>()),
+      name: map['name'] ?? '',
+      systemName: map['systemName'] ?? '',
+      systemVersion: map['systemVersion'] ?? '',
+      model: map['model'] ?? '',
+      localizedModel: map['localizedModel'] ?? '',
+      identifierForVendor: map['identifierForVendor'] ?? '',
+      isPhysicalDevice: map['isPhysicalDevice'] != null
+          ? map['isPhysicalDevice'] == 'true'
+          : false,
+      utsname: IosUtsname._fromMap(map['utsname'] != null
+          ? map['utsname'].cast<String, dynamic>()
+          : <String, dynamic>{}),
+      storage: Storage._fromMap(map['storage'] != null
+          ? map['storage'].cast<String, dynamic>()
+          : <String, dynamic>{}),
     );
   }
 
@@ -84,11 +90,11 @@ class IosDeviceInfo {
 /// See http://pubs.opengroup.org/onlinepubs/7908799/xsh/sysutsname.h.html for details.
 class IosUtsname {
   IosUtsname._({
-    this.sysname,
-    this.nodename,
-    this.release,
-    this.version,
-    this.machine,
+    required this.sysname,
+    required this.nodename,
+    required this.release,
+    required this.version,
+    required this.machine,
   });
 
   /// Operating system name.
@@ -109,11 +115,11 @@ class IosUtsname {
   /// Deserializes from the map message received from [_kChannel].
   static IosUtsname _fromMap(Map<String, dynamic> map) {
     return IosUtsname._(
-      sysname: map['sysname'],
-      nodename: map['nodename'],
-      release: map['release'],
-      version: map['version'],
-      machine: map['machine'],
+      sysname: map['sysname'] ?? '',
+      nodename: map['nodename'] ?? '',
+      release: map['release'] ?? '',
+      version: map['version'] ?? '',
+      machine: map['machine'] ?? '',
     );
   }
 
@@ -139,18 +145,18 @@ class Storage {
   num romAllB;
 
   Storage({
-    this.ramUseB,
-    this.ramAllB,
-    this.romUseB,
-    this.romAllB,
+    required this.ramUseB,
+    required this.ramAllB,
+    required this.romUseB,
+    required this.romAllB,
   });
 
   static Storage _fromMap(Map<String, dynamic> map) {
     return Storage(
-      ramUseB: map['ramUseB'],
-      ramAllB: map['ramAllB'],
-      romUseB: map['romUseB'],
-      romAllB: map['romAllB'],
+      ramUseB: map['ramUseB'] ?? 0,
+      ramAllB: map['ramAllB'] ?? 0,
+      romUseB: map['romUseB'] ?? 0,
+      romAllB: map['romAllB'] ?? 0,
     );
   }
 
